+++ plugins/org.preesm.model.scenario/src/org/preesm/model/scenario/util/ScenarioUserFactory.java
@@ -36,6 +36,7 @@ package org.preesm.model.scenario.util;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.preesm.model.scenario.EnergyConfig;
 import org.preesm.model.scenario.MemoryCopySpeedValue;
 import org.preesm.model.scenario.PapiEvent;
 import org.preesm.model.scenario.PapiEventModifier;
@@ -61,6 +62,15 @@ public class ScenarioUserFactory {
   }
 
   /**
+  *
+  */
+  public static final EnergyConfig createEnergyConfig() {
+    final EnergyConfig res = factory.createEnergyConfig();
+    res.setPerformanceObjective(factory.createPerformanceObjective());
+    return res;
+  }
+
+  /**
    *
    */
   public static final PapiEvent createTimingEvent() {
@@ -82,6 +92,7 @@ public class ScenarioUserFactory {
     createScenario.setPapifyConfig(createPapifyConfig());
     createScenario.setSimulationInfo(factory.createSimulationInfo());
     createScenario.setTimings(factory.createTimings());
+    createScenario.setEnergyConfig(createEnergyConfig());
     return createScenario;
   }